+++ settings/implementation_test.go
@@ -15,65 +15,102 @@ import (
 
 	"github.com/oscal-compass/oscal-sdk-go/extensions"
 	"github.com/oscal-compass/oscal-sdk-go/generators"
-	"github.com/oscal-compass/oscal-sdk-go/internal/set"
 )
 
-func TestGetFrameworkShortName(t *testing.T) {
+func TestMerge(t *testing.T) {
 	tests := []struct {
 		name                string
 		inputImplementation oscalTypes.ControlImplementationSet
-		wantName            string
-		wantFound           bool
+		assertFunc          func(t *testing.T, settings *ImplementationSettings)
 	}{
 		{
-			name: "Valid/ShortNameFromProps",
+			name: "Valid/ImplementationOnly",
 			inputImplementation: oscalTypes.ControlImplementationSet{
-				Props: &[]oscalTypes.Property{
+				SetParameters: &[]oscalTypes.SetParameter{
 					{
-						Name:  extensions.FrameworkProp,
-						Value: "propFramework",
-						Ns:    extensions.TrestleNameSpace,
+						ParamId: "my-test-param",
+						Values:  []string{"test-value"},
 					},
 				},
-				Source: "profiles/framework/profile.json",
+				ImplementedRequirements: []oscalTypes.ImplementedRequirementControlImplementation{
+					{
+						ControlId: "ex-1",
+						Props: &[]oscalTypes.Property{
+							{
+								Name:  extensions.RuleIdProp,
+								Value: "my-test-rule",
+								Ns:    extensions.TrestleNameSpace,
+							},
+						},
+					},
+				},
 			},
-			wantName:  "propFramework",
-			wantFound: true,
+			assertFunc: func(t *testing.T, impSettings *ImplementationSettings) {
+				settings := impSettings.AllSettings()
+				require.NotNil(t, settings)
+				require.True(t, settings.ContainsRule("my-test-rule"))
+				applicationControls, err := impSettings.ApplicableControls("my-test-rule")
+				require.NoError(t, err)
+				require.Len(t, applicationControls, 1)
+				require.Equal(t, applicationControls[0], "ex-1")
+				require.Contains(t, settings.selectedParameters, "my-test-param")
+			},
 		},
 		{
-			name: "Valid/ShortNameFromSource",
+			name: "Valid/ControlLevel",
 			inputImplementation: oscalTypes.ControlImplementationSet{
-				Source: "profiles/sourceFramework/profile.json",
+				ImplementedRequirements: []oscalTypes.ImplementedRequirementControlImplementation{
+					{
+						ControlId: "ex-1",
+						SetParameters: &[]oscalTypes.SetParameter{
+							{
+								ParamId: "my-test-param",
+								Values:  []string{"test-value"},
+							},
+						},
+						Props: &[]oscalTypes.Property{
+							{
+								Name:  extensions.RuleIdProp,
+								Value: "my-test-rule",
+								Ns:    extensions.TrestleNameSpace,
+							},
+						},
+					},
+				},
 			},
-			wantName:  "sourceFramework",
-			wantFound: true,
-		},
-		{
-			name:                "Valid/NoShortName",
-			inputImplementation: oscalTypes.ControlImplementationSet{},
-			wantName:            "",
-			wantFound:           false,
+			assertFunc: func(t *testing.T, impSettings *ImplementationSettings) {
+				settings := impSettings.AllSettings()
+				require.NotNil(t, settings)
+				require.True(t, settings.ContainsRule("my-test-rule"))
+				applicationControls, err := impSettings.ApplicableControls("my-test-rule")
+				require.NoError(t, err)
+				require.Len(t, applicationControls, 1)
+				require.Equal(t, applicationControls[0], "ex-1")
+				impRequirementSettings, err := impSettings.ByControlID("ex-1")
+				require.NoError(t, err)
+				require.Contains(t, impRequirementSettings.selectedParameters, "my-test-param")
+				require.True(t, impRequirementSettings.ContainsRule("my-test-rule"))
+			},
 		},
 	}
 
 	for _, c := range tests {
 		t.Run(c.name, func(t *testing.T) {
-			name, found := GetFrameworkShortName(c.inputImplementation)
-			require.Equal(t, c.wantName, name)
-			require.Equal(t, c.wantFound, found)
+			testSettings := prepSettings(t)
+			testSettings.merge(c.inputImplementation)
+			c.assertFunc(t, testSettings)
 		})
 	}
 }
 
-func TestFramework(t *testing.T) {
+func prepSettings(t *testing.T) *ImplementationSettings {
 	testDataPath := filepath.Join("../testdata", "component-definition-test-reqs.json")
+
 	file, err := os.Open(testDataPath)
 	require.NoError(t, err)
 	definition, err := generators.NewComponentDefinition(file)
 	require.NoError(t, err)
-
-	require.NotNil(t, definition.Components)
-
+	require.NotNil(t, definition)
 	var allImplementations []oscalTypes.ControlImplementationSet
 	for _, component := range *definition.Components {
 		if component.ControlImplementations == nil {
@@ -81,38 +118,7 @@ func TestFramework(t *testing.T) {
 		}
 		allImplementations = append(allImplementations, *component.ControlImplementations...)
 	}
-
-	implementationsMap, err := Framework("cis", allImplementations)
+	impSettings, err := Framework("cis", allImplementations)
 	require.NoError(t, err)
-	expectedSettings := &ImplementationSettings{
-		settings: RequirementSettings{
-			mappedRules: set.Set[string]{
-				"etcd_cert_file": struct{}{},
-				"etcd_key_file":  struct{}{},
-			},
-			selectedParameters: map[string]string{},
-		},
-		requirementSettings: map[string]RequirementSettings{
-			"CIS-2.1": {
-				mappedRules: set.Set[string]{
-					"etcd_cert_file": struct{}{},
-					"etcd_key_file":  struct{}{},
-				},
-				selectedParameters: map[string]string{},
-			},
-		},
-		controlsByRules: map[string]set.Set[string]{
-			"etcd_cert_file": {
-				"CIS-2.1": struct{}{},
-			},
-			"etcd_key_file": {
-				"CIS-2.1": struct{}{},
-			},
-		},
-	}
-
-	require.Equal(t, expectedSettings, implementationsMap)
-
-	_, err = Framework("doesnotexist", allImplementations)
-	require.EqualError(t, err, "framework doesnotexist is not in control implementations")
+	return impSettings
 }
