+++ settings/settings_test.go
@@ -11,6 +11,7 @@ import (
 	"sort"
 	"testing"
 
+	oscalTypes "github.com/defenseunicorns/go-oscal/src/types/oscal-1-1-2"
 	"github.com/stretchr/testify/require"
 
 	"github.com/oscal-compass/oscal-sdk-go/extensions"
@@ -18,10 +19,110 @@ import (
 	"github.com/oscal-compass/oscal-sdk-go/rules"
 )
 
+func TestNewSettings(t *testing.T) {
+	tests := []struct {
+		name             string
+		inputRequirement oscalTypes.ImplementedRequirementControlImplementation
+		wantSettings     Settings
+	}{
+		{
+			name: "Valid/MappedRulesFound",
+			inputRequirement: oscalTypes.ImplementedRequirementControlImplementation{
+				Props: &[]oscalTypes.Property{
+					{
+						Name:  extensions.RuleIdProp,
+						Ns:    extensions.TrestleNameSpace,
+						Value: "rule-1",
+					},
+					{
+						Name:  extensions.RuleIdProp,
+						Ns:    extensions.TrestleNameSpace,
+						Value: "rule-2",
+					},
+				},
+			},
+			wantSettings: Settings{
+				mappedRules: set.Set[string]{
+					"rule-1": struct{}{},
+					"rule-2": struct{}{},
+				},
+				selectedParameters: map[string]string{},
+			},
+		},
+		{
+			name: "Valid/ParametersFound",
+			inputRequirement: oscalTypes.ImplementedRequirementControlImplementation{
+				Props: &[]oscalTypes.Property{
+					{
+						Name:  extensions.RuleIdProp,
+						Ns:    extensions.TrestleNameSpace,
+						Value: "rule-1",
+					},
+					{
+						Name:  extensions.RuleIdProp,
+						Ns:    extensions.TrestleNameSpace,
+						Value: "rule-2",
+					},
+				},
+				SetParameters: &[]oscalTypes.SetParameter{
+					{
+						ParamId: "param-1",
+						Values: []string{
+							"value",
+						},
+					},
+				},
+			},
+			wantSettings: Settings{
+				mappedRules: set.Set[string]{
+					"rule-1": struct{}{},
+					"rule-2": struct{}{},
+				},
+				selectedParameters: map[string]string{
+					"param-1": "value",
+				},
+			},
+		},
+		{
+			name:             "Valid/NoSettingsFound",
+			inputRequirement: oscalTypes.ImplementedRequirementControlImplementation{},
+			wantSettings: Settings{
+				mappedRules:        map[string]struct{}{},
+				selectedParameters: map[string]string{},
+			},
+		},
+		{
+			name: "Invalid/MultipleParametersValues",
+			inputRequirement: oscalTypes.ImplementedRequirementControlImplementation{
+				SetParameters: &[]oscalTypes.SetParameter{
+					{
+						ParamId: "param-1",
+						Values: []string{
+							"value-1",
+							"value-2",
+						},
+					},
+				},
+			},
+			wantSettings: Settings{
+				mappedRules:        set.Set[string]{},
+				selectedParameters: map[string]string{},
+			},
+		},
+	}
+
+	for _, c := range tests {
+		t.Run(c.name, func(t *testing.T) {
+			gotSettings := NewSettingsFromImplementedRequirement(c.inputRequirement)
+			require.Equal(t, c.wantSettings, gotSettings)
+		})
+	}
+}
+
 func TestApplyToComponents(t *testing.T) {
 	tests := []struct {
 		name               string
-		settings           RequirementSettings
+		settings           Settings
 		componentID        string
 		expError           string
 		wantRules          []extensions.RuleSet
@@ -30,7 +131,7 @@ func TestApplyToComponents(t *testing.T) {
 		{
 			name:        "Valid/WithMappedRules",
 			componentID: "testComponent1",
-			settings: RequirementSettings{
+			settings: Settings{
 				mappedRules: set.Set[string]{
 					"testRule1": struct{}{},
 					"testRule2": struct{}{},
@@ -42,7 +143,7 @@ func TestApplyToComponents(t *testing.T) {
 		{
 			name:        "Valid/WithParameterOverrides",
 			componentID: "testComponent2",
-			settings: RequirementSettings{
+			settings: Settings{
 				selectedParameters: map[string]string{
 					"testParam1": "updatedValue",
 				},
@@ -79,7 +180,7 @@ func TestApplyToComponents(t *testing.T) {
 		{
 			name:        "Invalid/InvalidSettings",
 			componentID: "testComponent1",
-			settings: RequirementSettings{
+			settings: Settings{
 				mappedRules: set.Set[string]{
 					"doesnotexists": struct{}{},
 				},
@@ -93,7 +194,7 @@ func TestApplyToComponents(t *testing.T) {
 			testCtx := context.Background()
 			store := newFakeStore()
 
-			gotRules, err := ApplyToComponent(testCtx, c.componentID, store, &c.settings)
+			gotRules, err := ApplyToComponent(testCtx, c.componentID, store, c.settings)
 			sort.SliceStable(gotRules, func(i, j int) bool {
 				return gotRules[i].Rule.ID < gotRules[j].Rule.ID
 			})
