+++ transformers/doc.go
@@ -1,0 +1,8 @@
+/*
+ Copyright 2024 The OSCAL Compass Authors
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+// Package transformers generates OSCAL objects from others OSCAL objects to facilitate an
+// OSCAL-to-OSCAL transformation workflow.
+package transformers
